+++ CHANGELOG.md
@@ -1,4 +1,18 @@
-﻿## v0.1.2
+﻿## v0.3.0
+
+- Added loading element to the question page
+
+## v0.2.0
+
+- Added cookies token validation
+- Added landing page
+- Added register page
+- Created new menu logic
+- CSS Improvments
+- Added a Footer and a Header
+- Added Logout route
+
+## v0.1.2
 
 - Added "See password" button
 
@@ -10,6 +24,8 @@
 
 - Questions generator and validator implemented
 - Login implemented
+
+